edge/requirements.txt
requests
bcrypt
opencv-python
ultralytics
pynput

python-dotenv
